refactor(ATT-427): split firmware application god class into modular files
Break the 1360-line application.cpp into cohesive translation units so the
root file becomes a thin composition root, mirroring the ATT-424 api.cpp split.
Pure move of Application:: methods — zero behavior change.

- application.cpp (1360 -> 500): composition root — networkTask/ledTask,
  setup() callback wiring, loop()
- application_state.cpp: processState screen-routing state machine + updateLedState
- application_cards.cpp: processCardAuthenticationData NFC auth flow
- application_forms.cpp: resource usage form paging/render/submit flow
- application_session.cpp: resource/project selection, action buttons,
  pause accounting, disconnect reset

All four board envs build green (lite-ethernet exercises the non-LVGL path).
Bump FIRMWARE_VERSION 1.3.10 -> 1.3.11 for OTA.
