+++ apps/attractap/firmware/platformio.ini
@@ -17,7 +17,7 @@ build_type = debug
 extra_scripts = 
     pre:tools/build_adaptive_certs_wrapper.py
 build_flags =
-	-D FIRMWARE_VERSION='"1.3.10"'
+	-D FIRMWARE_VERSION='"1.3.11"'
 
 [env:attractap-touch]
 board = esp32-s3-devkitc-1-n16r8v
