+++ apps/attractap/firmware/src/application/application.cpp
@@ -1,3 +1,6 @@
+// Composition root: wires API/NFC/display callbacks, spawns tasks, runs loop
+// FEATURE: application
+
 #include "application.hpp"
 #include "../serial/serialCommandHandler.hpp"
 #ifdef ESP_PLATFORM
@@ -495,867 +498,3 @@ void Application::loop() {
   vTaskDelay(pdMS_TO_TICKS(1));
 #endif
 }
-
-void Application::processState() {
-#ifdef HAS_WS2812_LED
-  this->updateLedState();
-#endif
-
-  AttraccessApiConfig attraccessApiConfig = Settings::getAttraccessApiConfig();
-  bool connectionIsConfigured = !attraccessApiConfig.hostname.isEmpty() &&
-                                attraccessApiConfig.hostname != "" &&
-                                attraccessApiConfig.port > 0;
-
-    if (!connectionIsConfigured)
-    {
-        if (this->state != APPLICATION_STATE_CONFIGURATION_REQUIRED)
-        {
-            this->logger.debug("Connection not configured, showing config screen");
-            this->state = APPLICATION_STATE_CONFIGURATION_REQUIRED;
-
-#ifdef HAS_LVGL_DISPLAY
-      Display::connectionConfigurationScreen.disablePinLock();
-      Display::transitionToScreen(&Display::connectionConfigurationScreen);
-#endif
-    }
-
-    return;
-  }
-
-    if (this->state == APPLICATION_STATE_CONFIGURATION_REQUIRED)
-    {
-        return;
-    }
-
-#ifdef HAS_LVGL_DISPLAY
-  if (!this->bootDone &&
-      millis() - this->bootTime > APPLICATION_BOOT_SCREEN_DURATION) {
-    this->logger.debug("Boot screen duration reached, hiding boot screen");
-    this->bootDone = true;
-  }
-
-  if (!this->bootDone) {
-    return;
-  }
-
-  bool pinIsSet = Settings::getDeviceConfig().passCode != "0000";
-  if (!pinIsSet) {
-    if (this->state == APPLICATION_STATE_PIN_NOT_SET) {
-      return;
-    }
-
-    this->logger.debug("PIN is not set, showing pin screen");
-    this->state = APPLICATION_STATE_PIN_NOT_SET;
-
-    Display::transitionToScreen(&Display::setPinScreen);
-    return;
-  }
-#endif
-
-  State::ApiState apiState = State::getApiState();
-  State::NetworkState networkState = State::getNetworkState();
-  State::WebsocketState websocketState = State::getWebsocketState();
-  if (!apiState.authenticated ||
-      (!networkState.ethernet_connected && !networkState.wifi_connected) ||
-      !websocketState.connected) {
-#ifdef HAS_LVGL_DISPLAY
-    this->resetSessionOnDisconnect();
-#endif
-        if (this->state == APPLICATION_STATE_INIT)
-        {
-            return;
-        }
-
-        // User intentionally opened settings from the init screen — don't force back to init.
-        if (this->state == APPLICATION_STATE_CONFIGURATION_REQUIRED)
-        {
-            return;
-        }
-
-    this->logger.debug(
-        "API state is not authenticated, network state is not connected, "
-        "websocket state is not connected, showing init screen");
-    this->state = APPLICATION_STATE_INIT;
-
-#ifdef HAS_LVGL_DISPLAY
-    Display::transitionToScreen(&Display::initScreen);
-#endif
-    return;
-  }
-
-#ifdef HAS_LVGL_DISPLAY
-  if (this->externalState ==
-      EXTERNAL_STATE_ENROLL_NEW_CARD_GET_AVAILABLE_KEY_NO) {
-    if (this->state == APPLICATION_STATE_ENROLLMENT) {
-      uint32_t now = millis();
-      if (now - this->apiEnrollNewCardGetAvailableKeyNoStartTimeMs > 30000) {
-        this->logger.error(
-            "Enroll new card get available key number timeout reached");
-        this->externalState = EXTERNAL_STATE_NONE;
-        return;
-      }
-
-      uint8_t uid[7] = {0};
-      uint8_t uidLength = 0;
-      uint8_t keyNo = 0;
-      bool success = this->nfc.getAvailableKeyNo(uid, &uidLength, &keyNo);
-
-      if (success) {
-        this->api.sendEnrollNewCardAvailableKeyNo(uid, uidLength, keyNo);
-        this->externalState = EXTERNAL_STATE_NONE;
-      }
-      return;
-    }
-
-    this->nfc.disableCardDetection();
-#ifdef HAS_LVGL_DISPLAY
-    Display::enrollmentScreen.setUserName(
-        this->apiEnrollNewCardGetAvailableKeyNoData.username);
-#endif
-    this->apiEnrollNewCardGetAvailableKeyNoStartTimeMs = millis();
-#ifdef HAS_LVGL_DISPLAY
-    Display::enrollmentScreen.setEnrollmentTimeoutTime(
-        this->apiEnrollNewCardGetAvailableKeyNoStartTimeMs + 30000);
-    Display::transitionToScreen(&Display::enrollmentScreen);
-#endif
-
-    this->state = APPLICATION_STATE_ENROLLMENT;
-
-    return;
-  }
-
-  if (this->externalState == EXTERNAL_STATE_ENROLL_NEW_CARD) {
-    if (this->state == APPLICATION_STATE_ENROLLMENT) {
-      return;
-    }
-
-    this->nfc.enableCardDetection();
-    this->state = APPLICATION_STATE_ENROLLMENT;
-    return;
-  }
-#endif
-
-#ifndef HAS_LVGL_DISPLAY
-  if (this->cardDetected && !this->cardRemoved) {
-    unsigned long currentPresentationDurationMs =
-        millis() - this->cardDetectionTimeMs;
-    if (currentPresentationDurationMs > NFC_CARD_LONG_PRESENTATION_TIME_MS &&
-        !this->cardPresentationWasLong) {
-      this->beeper.indicateBeep();
-#ifdef HAS_WS2812_LED
-      this->led.triggerIndicate();
-#endif
-      this->cardPresentationWasLong = true;
-    }
-  }
-#endif
-
-  if (this->externalState == EXTERNAL_STATE_AUTHENTICATE_CARD) {
-    if (this->state == APPLICATION_STATE_AUTHENTICATE_CARD) {
-      return;
-    }
-
-#ifdef HAS_LVGL_DISPLAY
-    Display::resourceDetailsScreen.setUserDetails(
-        ResourceDetailsScreen::UserDetails{
-            .username = this->cardAuthenticationData.username,
-            .canManageResource = this->cardAuthenticationData.canManageResource,
-            .hasIntroduction = this->cardAuthenticationData.hasIntroduction,
-            .isIntroducer = this->cardAuthenticationData.isIntroducer});
-#endif
-
-    this->state = APPLICATION_STATE_AUTHENTICATE_CARD;
-
-#ifndef HAS_LVGL_DISPLAY
-    // For non-display mode, process authentication immediately since the card
-    // is still present and won't trigger another detection event
-    this->processCardAuthenticationData();
-#else
-    this->nfc.enableCardDetection();
-#endif
-    return;
-  }
-
-  if (this->externalState == EXTERNAL_STATE_FIRMWARE_UPDATE) {
-    if (this->state == APPLICATION_STATE_FIRMWARE_UPDATE) {
-#ifdef HAS_LVGL_DISPLAY
-      Display::firmwareUpdateScreen.setProgress(
-          this->firmwareUpdateProgressPct);
-      Display::firmwareUpdateScreen.setAvailableVersion(
-          this->availableFirmwareVersion);
-#endif
-      return;
-    }
-
-#ifdef HAS_LVGL_DISPLAY
-    Display::transitionToScreen(&Display::firmwareUpdateScreen);
-#endif
-    this->state = APPLICATION_STATE_FIRMWARE_UPDATE;
-#ifdef HAS_WS2812_LED
-    this->updateLedState();
-#endif
-    return;
-  }
-
-#ifdef HAS_LVGL_DISPLAY
-  if (this->resourceCount == 0) {
-    if (this->state == APPLICATION_STATE_NO_RESOURCES) {
-      return;
-    }
-
-    this->logger.debug("Resource count is 0, showing no resources screen");
-    this->state = APPLICATION_STATE_NO_RESOURCES;
-
-    Display::transitionToScreen(&Display::noResourcesScreen);
-    return;
-  }
-
-  if (this->resourceCount == 1 && !this->resourceIsSelected) {
-    this->logger.debug(
-        "Resource count is 1 and resource is not selected, selecting resource");
-    this->selectResource(resourceList.items[0]);
-    return;
-  }
-
-  if (this->resourceCount > 0 && !this->resourceIsSelected) {
-    if (this->resourceListUpdated) {
-// Update UI with the list
-#ifdef HAS_LVGL_DISPLAY
-      Display::resourceListScreen.setResourceList(this->resourceList);
-#endif
-      this->resourceListUpdated = false;
-    }
-
-    if (this->state == APPLICATION_STATE_RESOURCE_LIST) {
-      return;
-    }
-
-    this->logger.debug("Resource count is greater than 0 and resource is not "
-                       "selected, showing resource list");
-    this->state = APPLICATION_STATE_RESOURCE_LIST;
-#ifdef HAS_LVGL_DISPLAY
-    Display::transitionToScreen(&Display::resourceListScreen);
-#endif
-    return;
-  }
-
-  if (this->selectedResourceChanged) {
-    for (uint16_t i = 0; i < this->resourceList.count; ++i) {
-      if (this->resourceList.items[i].id == this->selectedResourceId) {
-        API::ResourceBrief resource = this->resourceList.items[i];
-
-        Display::lockscreen.setResourceName(resource.name);
-        Display::lockscreen.setUsageInfo(resource.hasActiveUsage,
-                                         resource.activeUser);
-
-        // Directly pass the native struct to the screen so it can avoid String
-        // conversions
-        Display::resourceDetailsScreen.setResourceAndUsageDetails(resource);
-
-        break;
-      }
-    }
-    this->selectedResourceChanged = false;
-  }
-
-  uint32_t now = millis();
-  if (!this->unlocked) {
-    if (this->state == APPLICATION_STATE_LOCKED) {
-
-      if (now - this->timeOfResourceSelectionMs >
-          this->RESOURCE_SELECTION_TIMEOUT_MS) {
-        this->logger.debug(
-            "Resource selection timeout reached, showing resource list");
-        this->resourceIsSelected = false;
-      }
-      return;
-    }
-
-    this->logger.debug("Card is not detected, showing lockscreen");
-    this->state = APPLICATION_STATE_LOCKED;
-#ifdef HAS_LVGL_DISPLAY
-    Display::transitionToScreen(&Display::lockscreen, [this]() {
-      this->logger.debug(
-          "Lockscreen transition complete, enabling card detection");
-      this->nfc.enableCardDetection();
-    });
-#else
-    this->nfc.enableCardDetection();
-#endif
-    return;
-  }
-
-  if (this->state == APPLICATION_STATE_UNLOCKED) {
-    // Subtract any accumulated pause time while actions were in-progress
-    uint32_t effectiveElapsed = now - this->timeOfUnlockedMs;
-    if (effectiveElapsed > this->accumulatedPauseMs) {
-      effectiveElapsed -= this->accumulatedPauseMs;
-    } else {
-      effectiveElapsed = 0;
-    }
-    if (effectiveElapsed > this->UNLOCKED_TIMEOUT_MS) {
-      this->logger.debug("Unlocked timeout reached, locking");
-      this->unlocked = false;
-      this->resourceIsSelected = this->resourceCount == 1;
-    }
-
-    if (this->projectsOfUserResponseUpdated) {
-#ifdef HAS_LVGL_DISPLAY
-      Display::resourceDetailsScreen.setProjects(this->projectsOfUserResponse);
-      Display::resourceDetailsScreen.setSelectedProject(
-          this->selectedProjectId, this->selectedProjectName.c_str());
-#endif
-      this->projectsOfUserResponseUpdated = false;
-    }
-
-    return;
-  }
-
-  this->logger.debug("Resource is unlocked, showing resource details screen");
-  this->state = APPLICATION_STATE_UNLOCKED;
-  this->restartSessionTimeout();
-
-  Display::transitionToScreen(&Display::resourceDetailsScreen);
-#else
-
-  // Process unlocked card actions for non-display mode
-  if (this->state == APPLICATION_STATE_AUTHENTICATE_CARD) {
-    if (!this->cardDetected) {
-      return;
-    }
-
-    if (!this->unlocked) {
-      return;
-    }
-
-    if (!this->cardRemoved) {
-      return;
-    }
-
-    this->logger.debug("Card detected and removed and unlocked, processing");
-
-    // Reset state flags before triggering action
-    this->unlocked = false;
-    this->cardDetected = false;
-    this->cardRemoved = false;
-
-    if (this->resourceIsDoor) {
-      if (this->cardPresentationWasLong) {
-        this->api.lockDoor(this->selectedResourceId);
-      } else {
-        this->api.unlockDoor(this->selectedResourceId);
-      }
-    } else {
-      if (this->cardPresentationWasLong) {
-        this->api.stopResourceUsageSession(this->selectedResourceId);
-      } else {
-        this->api.startResourceUsageSession(this->selectedResourceId);
-      }
-    }
-
-    // Reset state back to waiting for card
-    this->state = APPLICATION_STATE_WAIT_FOR_CARD;
-    this->nfc.enableCardDetection();
-    return;
-  }
-
-  if (this->state != APPLICATION_STATE_WAIT_FOR_CARD) {
-    this->logger.debug("Waiting for card detection");
-    this->state = APPLICATION_STATE_WAIT_FOR_CARD;
-    this->nfc.enableCardDetection();
-    return;
-  }
-#endif
-}
-
-#ifdef HAS_LVGL_DISPLAY
-void Application::handleConnectionConfigurationSave(
-    const ConnectionConfigurationScreen::ConnectionConfig &cfg) {
-  // split cfg.host into hostname and port (if no port present, use 443)
-  String hostname = cfg.host;
-  String port = "443";
-  if (cfg.host.indexOf(":") != -1) {
-    hostname = cfg.host.substring(0, cfg.host.indexOf(":"));
-    port = cfg.host.substring(cfg.host.indexOf(":") + 1);
-  }
-  Settings::saveNetworkConfig(cfg.ssid, cfg.password);
-  Settings::saveAttraccessApiConfig(hostname, port.toInt(), cfg.useSSL);
-
-    Settings::setDevicePin(cfg.devicePin);
-    Settings::setBeeperEnabled(cfg.beeperEnabled);
-
-    this->state = APPLICATION_STATE_INIT;
-    this->api.enableConnectionAttempts();
-    Display::transitionToScreen(&Display::initScreen);
-};
-
-void Application::handleResourceListUpdate(
-    const API::ResourceList &resourceList) {
-  this->logger.infof("Resource list updated: %d resources", resourceList.count);
-
-  this->resourceList = resourceList;
-  this->resourceCount = resourceList.count;
-  this->resourceListUpdated = true;
-
-  // If a resource is already selected, try to find it in the new list and
-  // refresh the details screen.
-  if (this->resourceIsSelected) {
-    this->logger.info(
-        "Resource is selected, trying to find it in the new list");
-    for (uint16_t i = 0; i < this->resourceList.count; ++i) {
-      const auto &obj = this->resourceList.items[i];
-      if (obj.id == this->selectedResourceId) {
-        this->logger.infof(
-            "Resource found in the new list, refreshing the details screen: %s",
-            obj.name);
-        this->selectResource(obj);
-        break;
-      }
-    }
-  }
-}
-#endif
-
-#ifdef HAS_WS2812_LED
-void Application::updateLedState() {
-  LedController::LedState ledState;
-  switch (this->state) {
-  case APPLICATION_STATE_CONFIGURATION_REQUIRED:
-    ledState = LedController::LED_STATE_CONFIG_REQUIRED;
-    break;
-  case APPLICATION_STATE_INIT:
-    ledState = LedController::LED_STATE_INIT;
-    break;
-  case APPLICATION_STATE_AUTHENTICATE_CARD:
-    ledState = LedController::LED_STATE_AUTHENTICATE_CARD;
-    break;
-  case APPLICATION_STATE_NO_RESOURCES:
-    ledState = LedController::LED_STATE_NO_RESOURCES;
-    break;
-  case APPLICATION_STATE_WAIT_FOR_CARD:
-    ledState = LedController::LED_STATE_WAIT_FOR_CARD;
-    break;
-  case APPLICATION_STATE_FIRMWARE_UPDATE:
-    ledState = LedController::LED_STATE_FIRMWARE_UPDATE;
-    break;
-  default:
-    ledState = LedController::LED_STATE_WAIT_FOR_CARD;
-    break;
-  }
-  this->led.setState(ledState);
-}
-#endif
-
-void Application::processCardAuthenticationData() {
-  this->logger.infof("Trying to authenticate with keyNo: %u",
-                     this->cardAuthenticationData.keyNo);
-  if (this->cardAuthenticationData.keyLen != 16) {
-    this->logger.error("Invalid key bytes provided");
-    this->beeper.errorBeep();
-#ifdef HAS_WS2812_LED
-    this->led.triggerError();
-#endif
-    this->nfc.enableCardDetection();
-    this->externalState = EXTERNAL_STATE_AUTHENTICATE_CARD;
-    return;
-  }
-
-  bool authenticated =
-      this->nfc.authenticate(this->cardAuthenticationData.keyNo,
-                             this->cardAuthenticationData.keyBytes);
-
-  if (!authenticated) {
-    this->logger.error("Authentication failed");
-    this->beeper.errorBeep();
-#ifdef HAS_WS2812_LED
-    this->led.triggerError();
-#endif
-    this->nfc.enableCardDetection();
-    this->externalState = EXTERNAL_STATE_AUTHENTICATE_CARD;
-    return;
-  }
-
-  this->beeper.successBeep();
-#ifdef HAS_WS2812_LED
-  this->led.triggerSuccess();
-#endif
-  this->logger.info("Authentication successful");
-
-  this->externalState = EXTERNAL_STATE_NONE;
-
-  this->unlocked = true;
-
-#ifndef HAS_LVGL_DISPLAY
-  // Enable card detection to detect card removal in non-display mode
-  this->nfc.enableCardDetection();
-#endif
-}
-
-#ifdef HAS_LVGL_DISPLAY
-void Application::selectResource(const API::ResourceBrief &resource) {
-  this->logger.infof("Resource selected: %s", resource.name);
-  this->resourceIsSelected = true;
-  this->selectedResourceId = resource.id;
-  this->restartResourceSelectionTimeout();
-  this->selectedResourceChanged = true;
-}
-
-void Application::requestProjectsPage(uint32_t page) {
-  if (page == 0) {
-    page = 1;
-  }
-  this->api.requestProjectsOfUser(page);
-}
-
-void Application::clearProjectSelection() {
-  this->selectedProjectId = 0;
-  this->selectedProjectName = "";
-  this->projectsCurrentPage = 1;
-  this->projectsTotalCount = 0;
-  this->projectsHasMore = false;
-  this->projectsOfUserResponse.count = 0;
-  this->projectsOfUserResponse.page = 1;
-  this->projectsOfUserResponse.total = 0;
-  this->projectsOfUserResponse.limit = API::MAX_PROJECTS_PER_PAGE;
-  this->projectsOfUserResponse.hasMore = false;
-  this->projectsOfUserResponseUpdated = true;
-  Display::resourceDetailsScreen.setSelectedProject(0, nullptr);
-}
-
-void Application::handleProjectSelection(uint32_t projectId,
-                                         const String &projectName) {
-  this->selectedProjectId = projectId;
-  this->selectedProjectName = projectName;
-  Display::resourceDetailsScreen.setSelectedProject(projectId,
-                                                    projectName.c_str());
-}
-
-void Application::handleFormsRequest(
-    const API::ResourceUsageFormRequest &request) {
-  // 'request' aliases this->pendingFormRequest (filled by the callback).
-  (void)request;
-  this->hasPendingFormRequest = true;
-  this->formCursorFormIdx = 0;
-  this->formCursorOffset = 0;
-  Display::resourceDetailsScreen.hideActionProgress();
-  Display::resourceDetailsScreen.showFormsModal(this->pendingFormRequest);
-  this->requestCurrentFormField();
-}
-
-uint32_t Application::totalFormFields() const {
-  uint32_t total = 0;
-  for (uint8_t i = 0;
-       i < this->pendingFormRequest.formCount && i < API::MAX_FORMS_PER_REQUEST;
-       ++i) {
-    total += this->pendingFormRequest.forms[i].fieldCount;
-  }
-  return total;
-}
-
-uint32_t Application::globalFormFieldNumber() const {
-  uint32_t number = 0;
-  for (uint8_t i = 0;
-       i < this->formCursorFormIdx && i < API::MAX_FORMS_PER_REQUEST; ++i) {
-    number += this->pendingFormRequest.forms[i].fieldCount;
-  }
-  return number + this->formCursorOffset + 1;
-}
-
-bool Application::isLastFormField() const {
-  // Last when no later field exists in this or any subsequent form.
-  if (this->formCursorFormIdx >= this->pendingFormRequest.formCount) {
-    return true;
-  }
-  if (this->formCursorOffset + 1 <
-      this->pendingFormRequest.forms[this->formCursorFormIdx].fieldCount) {
-    return false;
-  }
-  for (uint8_t i = this->formCursorFormIdx + 1;
-       i < this->pendingFormRequest.formCount && i < API::MAX_FORMS_PER_REQUEST;
-       ++i) {
-    if (this->pendingFormRequest.forms[i].fieldCount > 0) {
-      return false;
-    }
-  }
-  return true;
-}
-
-void Application::requestCurrentFormField() {
-  // Skip forms that carry no fields, finish when the cursor runs past the end.
-  while (this->formCursorFormIdx < this->pendingFormRequest.formCount &&
-         this->pendingFormRequest.forms[this->formCursorFormIdx].fieldCount ==
-             0) {
-    this->formCursorFormIdx++;
-    this->formCursorOffset = 0;
-  }
-  if (this->formCursorFormIdx >= this->pendingFormRequest.formCount) {
-    this->finishFormFlow();
-    return;
-  }
-  const API::ResourceUsageFormMeta &form =
-      this->pendingFormRequest.forms[this->formCursorFormIdx];
-  this->api.requestFormFields(this->pendingFormRequest.resourceId,
-                              this->pendingFormRequest.action, form.id,
-                              this->formCursorOffset, API::MAX_FORM_PAGE_FIELDS);
-}
-
-void Application::advanceFormCursor() {
-  this->formCursorOffset += API::MAX_FORM_PAGE_FIELDS;
-  if (this->formCursorFormIdx < this->pendingFormRequest.formCount &&
-      this->formCursorOffset >=
-          this->pendingFormRequest.forms[this->formCursorFormIdx].fieldCount) {
-    this->formCursorFormIdx++;
-    this->formCursorOffset = 0;
-  }
-}
-
-void Application::retreatFormCursor() {
-  if (this->formCursorOffset > 0) {
-    this->formCursorOffset--;
-  } else {
-    if (this->formCursorFormIdx == 0) {
-      return;
-    }
-    int16_t idx = static_cast<int16_t>(this->formCursorFormIdx) - 1;
-    while (idx >= 0 && this->pendingFormRequest.forms[idx].fieldCount == 0) {
-      idx--;
-    }
-    if (idx < 0) {
-      return;
-    }
-    this->formCursorFormIdx = static_cast<uint8_t>(idx);
-    this->formCursorOffset = this->pendingFormRequest.forms[idx].fieldCount - 1;
-  }
-  this->requestCurrentFormField();
-}
-
-void Application::handleFormFields(const API::ResourceUsageFormFieldsPage &page) {
-  if (!this->hasPendingFormRequest) {
-    return;
-  }
-  bool canGoBack = this->globalFormFieldNumber() > 1;
-  Display::resourceDetailsScreen.renderFormField(
-      page, canGoBack, this->isLastFormField(), this->globalFormFieldNumber(),
-      this->totalFormFields());
-}
-
-void Application::handleFormPageNext(const API::FormPageSubmission &page) {
-  if (!this->hasPendingFormRequest) {
-    return;
-  }
-  this->api.submitFormPage(this->pendingFormRequest.resourceId,
-                           this->pendingFormRequest.action, page);
-}
-
-void Application::handleFormPageResult(
-    const API::ResourceUsageFormPageResult &result) {
-  if (!this->hasPendingFormRequest) {
-    return;
-  }
-  if (result.valid) {
-    this->advanceFormCursor();
-    this->requestCurrentFormField();
-  } else {
-    Display::resourceDetailsScreen.showFormPageErrors(result);
-  }
-}
-
-void Application::handleFormPageBack() {
-  if (!this->hasPendingFormRequest) {
-    return;
-  }
-  this->retreatFormCursor();
-}
-
-void Application::finishFormFlow() {
-  this->hasPendingFormRequest = false;
-  Display::resourceDetailsScreen.hideFormsModal();
-  Display::resourceDetailsScreen.showActionProgress("Sende Formular");
-
-  if (this->pendingActionType == PENDING_ACTION_START_SESSION) {
-    this->api.startResourceUsageSession(this->pendingActionResourceId,
-                                        this->pendingActionProjectId);
-  } else if (this->pendingActionType == PENDING_ACTION_STOP_SESSION) {
-    this->api.stopResourceUsageSession(this->pendingActionResourceId);
-  } else {
-    this->handleFormsCancel();
-  }
-}
-
-void Application::handleFormsCancel() {
-  if (!this->hasPendingFormRequest) {
-    return;
-  }
-  this->hasPendingFormRequest = false;
-  this->pendingActionType = PENDING_ACTION_NONE;
-  this->formCursorFormIdx = 0;
-  this->formCursorOffset = 0;
-  Display::resourceDetailsScreen.hideFormsModal();
-  Display::resourceDetailsScreen.hideActionProgress();
-  this->endActionPause();
-}
-
-void Application::onActionResult(const String &eventType) {
-  if (eventType == "START_RESOURCE_USAGE_SESSION" ||
-      eventType == "STOP_RESOURCE_USAGE_SESSION") {
-    this->pendingActionType = PENDING_ACTION_NONE;
-    this->hasPendingFormRequest = false;
-    Display::resourceDetailsScreen.hideFormsModal();
-  }
-}
-
-void Application::handleTouch(int16_t x, int16_t y) {
-  if (this->state == APPLICATION_STATE_UNLOCKED) {
-    this->restartSessionTimeout();
-  }
-}
-
-void Application::restartSessionTimeout() {
-  uint32_t now = millis();
-  Display::resourceDetailsScreen.setSessionTimeoutTime(
-      now + this->UNLOCKED_TIMEOUT_MS);
-  this->timeOfUnlockedMs = now;
-  this->resetPauseAccounting();
-}
-
-void Application::handleResourceDetailsButtonClick(
-    ResourceDetailsScreen::ButtonClickEventData evt) {
-  this->logger.infof("Resource details button clicked: %d",
-                     evt.buttonClickType);
-
-  if (this->state != APPLICATION_STATE_UNLOCKED) {
-    return;
-  }
-
-  switch (evt.buttonClickType) {
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_START_SESSION:
-    Display::resourceDetailsScreen.showActionProgress("Starte Sitzung");
-    this->beginActionPause();
-    this->pendingActionType = PENDING_ACTION_START_SESSION;
-    this->pendingActionResourceId = this->selectedResourceId;
-    this->pendingActionProjectId = this->selectedProjectId;
-    this->hasPendingFormRequest = false;
-    this->api.startResourceUsageSession(this->selectedResourceId,
-                                        this->selectedProjectId);
-    break;
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_STOP_SESSION:
-    Display::resourceDetailsScreen.showActionProgress("Beende Sitzung");
-    this->beginActionPause();
-    this->pendingActionType = PENDING_ACTION_STOP_SESSION;
-    this->pendingActionResourceId = this->selectedResourceId;
-    this->pendingActionProjectId = 0;
-    this->hasPendingFormRequest = false;
-    this->api.stopResourceUsageSession(this->selectedResourceId);
-    break;
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_LOCK_DOOR:
-    Display::resourceDetailsScreen.showActionProgress("Sperre Tuer");
-    this->beginActionPause();
-    this->api.lockDoor(this->selectedResourceId);
-    break;
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_UNLOCK_DOOR:
-    Display::resourceDetailsScreen.showActionProgress("Entsperre Tuer");
-    this->beginActionPause();
-    this->api.unlockDoor(this->selectedResourceId);
-    break;
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_UNLATCH_DOOR:
-    Display::resourceDetailsScreen.showActionProgress("Oeffne Tuer-Riegel");
-    this->beginActionPause();
-    this->api.unlatchDoor(this->selectedResourceId);
-    break;
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_FLOW_BUTTON:
-    Display::resourceDetailsScreen.showActionProgress("Aktion Ausfuehren");
-    this->beginActionPause();
-    this->api.triggerFlowButton(this->selectedResourceId, evt.flowButtonId);
-    break;
-  case ResourceDetailsScreen::BUTTON_CLICK_TYPE_LOGOUT:
-    if (this->resourceCount > 1) {
-      this->resourceIsSelected = false;
-    }
-    this->unlocked = false;
-    this->currentProjectsUser = "";
-    this->clearProjectSelection();
-    this->pendingActionType = PENDING_ACTION_NONE;
-    this->hasPendingFormRequest = false;
-    Display::resourceDetailsScreen.hideFormsModal();
-    break;
-  }
-}
-
-void Application::restartResourceSelectionTimeout() {
-  uint32_t now = millis();
-  this->timeOfResourceSelectionMs = now;
-}
-
-void Application::beginActionPause() {
-  this->actionInProgressCount++;
-  if (this->actionInProgressCount == 1) {
-    this->pauseStartMs = millis();
-    // Freeze the UI indicator
-    Display::resourceDetailsScreen.setSessionTimeoutPaused(true);
-  }
-}
-
-void Application::endActionPause() {
-  if (this->actionInProgressCount == 0) {
-    return;
-  }
-  this->actionInProgressCount--;
-  if (this->actionInProgressCount == 0) {
-    uint32_t now = millis();
-    uint32_t delta =
-        (now >= this->pauseStartMs) ? (now - this->pauseStartMs) : 0;
-    this->accumulatedPauseMs += delta;
-    // Extend the UI deadline by the same delta and unfreeze
-    Display::resourceDetailsScreen.extendSessionTimeoutBy(delta);
-    Display::resourceDetailsScreen.setSessionTimeoutPaused(false);
-  }
-}
-
-void Application::resetPauseAccounting() {
-  this->pauseStartMs = 0;
-  this->accumulatedPauseMs = 0;
-  this->actionInProgressCount = 0;
-  // Ensure not paused visually
-  Display::resourceDetailsScreen.setSessionTimeoutPaused(false);
-}
-
-void Application::resetSessionOnDisconnect() {
-  bool sessionActive = this->unlocked || this->resourceIsSelected ||
-                       this->pendingActionType != PENDING_ACTION_NONE ||
-                       this->hasPendingFormRequest ||
-                       this->pendingFormRequestReady ||
-                       this->currentProjectsUser.length() > 0;
-
-  if (!sessionActive) {
-    return;
-  }
-
-  this->logger.info("Connectivity lost; resetting session state");
-
-  // Ensure any in-progress UI overlays are dismissed
-  Display::resourceDetailsScreen.hideActionProgress();
-  Display::resourceDetailsScreen.hideFormsModal();
-  this->resetPauseAccounting();
-
-  this->pendingActionType = PENDING_ACTION_NONE;
-  this->pendingActionResourceId = 0;
-  this->pendingActionProjectId = 0;
-  this->hasPendingFormRequest = false;
-  this->pendingFormRequestReady = false;
-  this->pendingFormFieldsReady = false;
-  this->pendingFormPageResultReady = false;
-  this->formCursorFormIdx = 0;
-  this->formCursorOffset = 0;
-
-  this->clearProjectSelection();
-  this->currentProjectsUser = "";
-
-  this->selectedResourceId = 0;
-  this->resourceIsSelected = false;
-  this->selectedResourceChanged = false;
-
-  this->unlocked = false;
-  this->externalState = EXTERNAL_STATE_NONE;
-  this->nfc.enableCardDetection();
-}
-#endif
